+++ src/features/comment/compose/reply/CommentReply.tsx
@@ -3,7 +3,6 @@ import {
   IonButtons,
   IonButton,
   IonHeader,
-  IonContent,
   IonToolbar,
   IonTitle,
   useIonToast,
@@ -15,53 +14,14 @@ import {
   PersonMentionView,
   PostView,
 } from "lemmy-js-client";
-import { useEffect, useRef, useState } from "react";
+import { useEffect, useState } from "react";
 import ItemReplyingTo from "./ItemReplyingTo";
-import useClient from "../../../helpers/useClient";
-import { useAppDispatch, useAppSelector } from "../../../store";
-import { Centered, Spinner } from "../../auth/Login";
-import { handleSelector, jwtSelector } from "../../auth/authSlice";
-import { css } from "@emotion/react";
-import { preventPhotoswipeGalleryFocusTrap } from "../../gallery/GalleryImg";
-import TextareaAutosizedForOnScreenKeyboard from "../../shared/TextareaAutosizedForOnScreenKeyboard";
-import { receivedComments } from "../commentSlice";
-import MarkdownToolbar, {
-  TOOLBAR_HEIGHT,
-  TOOLBAR_TARGET_ID,
-} from "../../shared/markdown/editing/MarkdownToolbar";
-import useKeyboardHeight from "../../../helpers/useKeyboardHeight";
-
-export const Container = styled.div<{ keyboardHeight: number }>`
-  min-height: 100%;
-
-  display: flex;
-  flex-direction: column;
-
-  padding-bottom: ${({ keyboardHeight }) =>
-    keyboardHeight
-      ? TOOLBAR_HEIGHT
-      : `calc(${TOOLBAR_HEIGHT} + env(safe-area-inset-bottom))`};
-`;
-
-export const Textarea = styled(TextareaAutosizedForOnScreenKeyboard)`
-  border: 0;
-  background: none;
-  resize: none;
-  outline: 0;
-  padding: 1rem;
-
-  min-height: 200px;
-
-  flex: 1 0 auto;
-
-  ${({ theme }) =>
-    !theme.dark &&
-    css`
-      .ios & {
-        background: var(--ion-item-background);
-      }
-    `}
-`;
+import useClient from "../../../../helpers/useClient";
+import { useAppDispatch, useAppSelector } from "../../../../store";
+import { Centered, Spinner } from "../../../auth/Login";
+import { handleSelector, jwtSelector } from "../../../auth/authSlice";
+import { receivedComments } from "../../commentSlice";
+import CommentContent from "../shared";
 
 export const UsernameIonText = styled(IonText)`
   font-size: 0.7em;
@@ -84,6 +44,9 @@ type CommentReplyProps = {
   item: CommentReplyItem;
 };
 
+/**
+ * New comment replying to something
+ */
 export default function CommentReply({
   dismiss,
   setCanDismiss,
@@ -98,8 +61,6 @@ export default function CommentReply({
   const [present] = useIonToast();
   const [loading, setLoading] = useState(false);
   const userHandle = useAppSelector(handleSelector);
-  const keyboardHeight = useKeyboardHeight();
-  const textareaRef = useRef<HTMLTextAreaElement>(null);
 
   async function submit() {
     if (!jwt) return;
@@ -183,25 +144,10 @@ export default function CommentReply({
           </IonButtons>
         </IonToolbar>
       </IonHeader>
-      <IonContent {...preventPhotoswipeGalleryFocusTrap}>
-        <Container keyboardHeight={keyboardHeight}>
-          <Textarea
-            ref={textareaRef}
-            value={replyContent}
-            onChange={(e) => setReplyContent(e.target.value)}
-            autoFocus
-            id={TOOLBAR_TARGET_ID}
-          />
-          <ItemReplyingTo item={item} />
-        </Container>
-      </IonContent>
 
-      <MarkdownToolbar
-        type="comment"
-        text={replyContent}
-        setText={setReplyContent}
-        textareaRef={textareaRef}
-      />
+      <CommentContent text={replyContent} setText={setReplyContent}>
+        <ItemReplyingTo item={item} />
+      </CommentContent>
     </>
   );
 }
